% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-transform.R
\name{arrange_inputs}
\alias{arrange_inputs}
\title{arrange_inputs(data, inputorder)}
\usage{
arrange_inputs(data, inputorder)
}
\arguments{
\item{data}{a dataframe containing the tracked sensor data for one subject}

\item{inputorder}{a named character vector, names needs to be Input1 to Input8, values only "EKG, "BVP", or "EDA"}
}
\value{
dataframe
}
\description{
Function order the inputs as they are needed for further physio processings. Additional, all Inputs
which are not part of the parameter inputorder will be removed.
}
\examples{
 
inputorder <-c(EKG = "Input1", BVP = "Input2", EDA ="Input3")
}

