% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{combineFiles}
\alias{combineFiles}
\title{combineFiles(files, type)}
\usage{
combineFiles(files, type)
}
\arguments{
\item{files;}{a vector}

\item{type;}{a character constant}
}
\value{
dataframe
}
\description{
Binds files of the same type in one dataframe and separate them by adding a colum SUBJETCT_ID_SUBJECT
}

