% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-compress_eda.R
\name{compress_eda.physio}
\alias{compress_eda.physio}
\title{compress_eda.physio()}
\usage{
compress_eda.physio(data, start_time)
}
\arguments{
\item{data;}{a physio dataframe for a specific subject}

\item{start_time;}{a POSIXct value, time when experiment started. It's usually the minimal time for subject in exp or web data.}
}
\value{
dataframe
}
\description{
Function compresses a physio dataframe by extracting only the EDA data ($Input2) and calculating the average eda values per second.
A start_time needs to be speficied to sync the data to a marker object.
}
\examples{
start_time <- min(labSession_web$Time, na.rm = T)
compress_eda.physio(data = physio_data, start_time)

}

