% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{convertTime}
\alias{convertTime}
\title{convertTime(data, type)}
\usage{
convertTime(data, type)
}
\arguments{
\item{data;}{a dataframe}

\item{type;}{a character constant}
}
\value{
dataframe
}
\description{
convertTime(data, type)
}

