% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-createUnisensData.R
\name{createUnisensData}
\alias{createUnisensData}
\title{createUnisensData}
\usage{
createUnisensData(data = physio174, subjectname = "174",
  destinationDirectory = "tests/testthat/datafortestingphysio/",
  nameUnisensDataFolder = "01_unisens_data",
  nameTrimmedUnisensDataFolder = "02_unisens_trimmed",
  pathMarkerFile = "tests/testthat/datafortestingphysio/marker174.csv")
}
\arguments{
\item{data}{a dataframe, pyhsio sensordata for one subject}

\item{subjectname}{a char name of a subject, for e.g. the subject_id_subject number. It will be the name of the subfolder with Unisense data.}

\item{destinationDirectory}{a char directory where all the files and subfolder will be placed at.}

\item{nameUnisensDataFolder}{a char name of a subdirectory in destinationDirectory where all the preproccessed unisens files will be created.}

\item{nameTrimmedUnisensDataFolder}{a char name of a subdirectory in destinationDirectory where all the final unisens files will be placed at.}

\item{pathMarkerFile}{a char, path to the marker file as csv. The function createMarkerFile will create a dataframe in the proper format.}
}
\value{
prints log files of processing and finally the path to all avaliable unisense data for each subject
}
\description{
This function groups a couple of other functions which create together Unisens data out of a MarkerFile and physio sensor data of one subject
}

