% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{detectDataType}
\alias{detectDataType}
\title{detectDataType(path)}
\usage{
detectDataType(file)
}
\arguments{
\item{path;}{a character}
}
\value{
constant web, physio, exp or none
}
\description{
Detect the data type of a given file
}

