% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{import(path, prefix)}
\usage{
import(path, prefix)
}
\arguments{
\item{path;}{a character}

\item{prefix;}{a character, nameing of lists}
}
\value{
none (adds list directly to global environment)
}
\description{
Imports brownie data (of heterogenous types) from folder and adds list containing dataframes and type to global environment (nameing by prefix).
It only imports data of class web and exp directly in R, due to the fact, that pyhsio-data is too big, it moves it in 
a subdirectory, in *path*/physio/
}

