% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{indexFiles}
\alias{indexFiles}
\title{indexFiles(path)}
\usage{
indexFiles(path)
}
\arguments{
\item{path;}{a charater}
}
\value{
array with file and corresponding datatype
}
\description{
Creates an index of  files in a folder with corresponding data types.
}

