% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCSVS.R
\name{load.csvs}
\alias{load.csvs}
\title{load.csvs}
\usage{
load.csvs()
}
\value{
This function loads all csv in the current working directory and saves them in a dataframe.
load.CSVS()
}
\description{
load.csvs
}
