% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-transform.R
\name{marker}
\alias{marker}
\title{marker(data, subject)}
\usage{
marker(data, subject, minSampleTimeSubject)
}
\arguments{
\item{data}{a list of class exp or web}

\item{subject, }{a integer identifies subject for which marker file will be created}

\item{minSampleTimeSubject, }{a integer - minimum sample time of subject is needed to caluclate durations of screens and sync sample to exp data}
}
\value{
a dataframe
}
\description{
Creates a "Marker File" for a specified subject. It's needed to calculate for e.g. heart reates for each experimental screen.
For unisens, this file needs to be in *wd*/*data*/physio .
}

