% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-transform_eda.R
\name{match_marker_eda.physio}
\alias{match_marker_eda.physio}
\title{match_marker_eda.physio()}
\usage{
match_marker_eda.physio(data, marker)
}
\arguments{
\item{data;}{a dataframe generated by compress_eda.physio()}

\item{marker;}{a dataframe generate by marker()}
}
\value{
dataframe
}
\description{
Matches marker object and compressed physio EDA data. Output can be plotted or used for further summary calculations.
}

