% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuroIS-package.R
\docType{package}
\name{NeuroIS-package}
\alias{NeuroIS-package}
\title{NeuroIS: The R-Package for Neuro Information Systems Research in R}
\description{
NeuroIS has become an established method in IS research for investigating user behavior, perception, 
and even neurophysiology. This same named package NeuroIS is a freely-available open source R-package 
for analyzing NeuroIS data, enabling scholars to investigate the bio-physiological fundations of 
human-computer interaction in individual as well as in group settings.
}
\details{
[1] http://im.iism.kit.edu/1093_1100.php

\tabular{ll}{ Package: \tab Brownie\cr Type: \tab Package\cr Version:
\tab 0.0.0.9000\cr Date: \tab 2017-31-01\cr License: \tab GPL-3\cr Depends: \tab
R (>= 3.0) \cr }
}
\examples{
### 0. Preparations & Import
library(NeuroIS)

import(path = "data/jump_and_rest", prefix = "jump_and_rest", 
       create_physio_object = T, physio_mv_dir = T)

### 1. Do central pages exist in terms of duration or click frequency?
summary(data = jump_and_rest_web, objectives = F)

### 2. At which pages did the proband jump?
marker <- marker(data = jump_and_rest_web, subject = 1, path_marker = "data/jump_and_rest/physio" )

start_time <- min(jump_and_rest_web[jump_and_rest_web$SUBJECT_ID_SUBJECT == 1]$Time, na.rm = T)

summary(jump_and_rest_physio, type = "eda", subject = 1, marker = marker, start_time = start_time) 

plot(jump_and_rest_physio, type = "eda", subject = 1, marker = marker, start_time)

plot(jump_and_rest_physio, type = "ecg", subject = 1, marker = "data/jump_and_rest/physio/marker1.csv",
     start_time, physio_unisens_dir = "data/jump_and_rest/physio/")

### 3. Which elements on a central page did the mouse last on?
web_summary <- summary(data = jump_and_rest_web, objectives = F)
urls <- web_summary$URL
take_screenshot(urls = urls)
plot(jump_and_rest_web, url = urls[4], type = "motion", subject = 1,
     alpha = 0.1, size = 3, color = "purple")

### 4. What’s the probability to go from one page to another, based on the subject?
library(clickstream)

cls <- as_clickstream(data = jump_and_rest_web, objectives = F, attribution ="last")
summary(cls)
mc <- fitMarkovChain(clickstreamList = cls, order = 1, control = list(optimizer = "quadratic"))
summary(mc)
plot(mc, order = 1)

}
\author{
Dominik Jung \email{d.jung@kit.edu}

Sven Michalczyk \email{sven.michalczyk@icloud.com}
}
\keyword{analytics,}
\keyword{brownie,}
\keyword{clickstream,}
\keyword{heatmap}
\keyword{kd2lab,}
\keyword{logfiles,}
\keyword{web}

