% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-transform.R
\name{normData}
\alias{normData}
\title{normData()}
\usage{
normData(data)
}
\arguments{
\item{data}{a physio dataframe}
}
\value{
dataframe
}
\description{
Normalize data colums using z score for colums with name contains "input".
}

