% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-plot.R
\name{plot.physio}
\alias{plot.physio}
\title{plot.physio()}
\usage{
\method{plot}{physio}(data, marker, type, subject, start_time,
  physio_unisens_dir)
}
\arguments{
\item{data;}{a list of class physio}

\item{type;}{a char, specify type of physio data you want to plot. Values are "eda" or "ecg". For type "ecg" it's important that
physio data is moved to folder *wd*/data/physio/}

\item{subject;}{a integer, specify subject for which summary should be shown}

\item{marker;}{a dataframe generate by marker() if type is "eda" or a path if type is "ecg"}

\item{start_time;}{a POSIXt, minimal time in exp or a web data for the subject which is specified}

\item{physio_unisens_dir;}{a path, directory where unisense data is created and contains physio csv-files
in the subject parameter of this function}
}
\value{
dataframe
}
\description{
Visualize physio data for one subject. Currently only plot of EDA and ECG data is possible.
}

