% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-plot.R
\name{plot.web}
\alias{plot.web}
\title{plot.web() Heatmap plotting on website screenshot}
\usage{
\method{plot}{web}(data, url, type = "motion", subject = c(1, 2),
  alpha = 0.1, size = 0.5, color = "purple", ...)
}
\arguments{
\item{data}{a list of class web}

\item{url}{a character,pPath to the website you want to plot data points at.
It requires that a screenshot is taken with take_screenshots of url.}

\item{type}{a character, "motion" or "click". Selects the type of data to plot.
"motion" Selects only mouse movements, "click" only mouse clicks.}

\item{subject}{a integer vector including the subjects which shall be plotted}

\item{alpha}{a numeric, From zero to 1; to set the transparency of data points.}

\item{size}{a numeric, For type motion size=0.5 is recommended and for type click size=2.5 is a good setting to start.
It depends very much on the amount of data points.}

\item{color}{a color, Color of data points. Use a color which is most different than the website. See the R's color specifications for more details and possible values.}
}
\value{
plot as png at folder "yourworkingdirectory/plots/plot.png"
}
\description{
Function is one of the core functionalitys of the package. You can plot either the mouse motion or the clicks
on a website screenshot. See the Argument description for customization details.
}
\examples{
plot(data, url="http://youwebsitetoplot.com/", type = "motion", subject = c(1, 2), alpha = 0.1, size = 0.5, color ="purple")
}

