% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-transform.R
\name{removeEmptyColums}
\alias{removeEmptyColums}
\title{removeEmptyColums()}
\usage{
removeEmptyColums(data)
}
\arguments{
\item{data}{a physio dataframe}
}
\value{
dataframe
}
\description{
Removes every colum containing zeros, as it appears for inputs with no connected sensor in a experiment.
}

