% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{rm_empty_col}
\alias{rm_empty_col}
\title{rm_empty_col()}
\usage{
rm_empty_col(data)
}
\arguments{
\item{data}{a physio dataframe}
}
\value{
dataframe
}
\description{
Removes every colum containing zeros, as it appears for inputs with no connected sensor in a experiment.
This reduces the size to a half.
}

