% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-summary.R
\name{summary.physio}
\alias{summary.physio}
\title{summary.physio()}
\usage{
\method{summary}{physio}(data, type = "eda", subject, marker, start_time)
}
\arguments{
\item{data;}{a list of class physio}

\item{type;}{a char, specify type of data you want to summarize. Currently only a summary of EDA is possible. BVG and ECG are planned.}

\item{subject;}{a integer, specify subject for which summary should be shown}

\item{marker;}{a dataframe generate by marker()}

\item{start_time;}{a POSIXt, minimal time in exp or a web data for the subject which is specified in the subject parameter of this function}
}
\value{
dataframe
}
\description{
Implemented as additional method for the gerneic function summary() and objects of class "physio". 
Run this function to get average EDA values per Screen or visited web pages.
passed to the argument objectives.
}

