% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-summary.R
\name{summary.web}
\alias{summary.web}
\title{Web Summary}
\usage{
\method{summary}{web}(data, objectives = F, ...)
}
\arguments{
\item{data;}{a list of class web}

\item{objectives;}{can be either FALSE or a labeled vector of URLs}
}
\value{
dataframe
}
\description{
Implemented as additional method for the gerneic function summary() and all objects of class "web". Run this function to get a first overview of the behaviour of the propants during a web experiment.
Create a summary of type "click frequencies" for objectives equals FALSE or "objectives" for a vector
passed to the argument objectives.
}
\details{
Summary "click frequency" has the following colums: URL, Sessions, Impressions, Impressions per Session, Duration.
Summary "objectives" has the colums Label, URL, Unique_Coversion, All_Conversions, Unique_Conversion-Rate,
and All_Conversion-Rate.
These metrics are an implemention of the industry standard in web analytics. In other words Google Analytics,
Piwik or Webtrekk for example use the same calculation techniques:
\itemize{
 \item{A Session is a unique page impression (URL call) of a propant (user) during an experiment (website visit)}
 \item{Impressions is the total number of page impressions}
 \item{Impressions per Visit is the average number of impressions per session; }
 \item{Duration is the average duration of a website visist per impression}
 \item{Unique Conversions is the number of sessions of the defined objective pages}
 \item{All Conversions is the number of impressions of the defined objective pages}
 \item{Unique Conversion-Rate is calculated by Unique Conversions divided by total number of impressions for all pages}
 \item{All Conversion-Rate is calculated by All Conversions divided by total number of impressions for all pages}
}
}
\examples{
objectives <- c(Fonds = "https://www.visualvest.de/fonds/",
                 Depot = "https://anlegen.visualvest.de/app/depot/")
summary(data = data, objectives = objectives)
}

