% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-summary_eda.R
\name{summary_eda.physio}
\alias{summary_eda.physio}
\title{summary_eda.physio()}
\usage{
summary_eda.physio(data)
}
\arguments{
\item{data;}{a dataframe generated by match_marker_eda.physio}
}
\value{
dataframe
}
\description{
Calculates the average eda values per event. Arrousels per Event can be identified. Time information is kept as well.
}

