% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-summary.R
\name{summary_web_imp}
\alias{summary_web_imp}
\title{Web Impression Summary}
\usage{
summary_web_imp(data)
}
\arguments{
\item{data, }{dataframe with web data}
}
\value{
dataframe
}
\description{
Web Impression Summary
}

