% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-summary.R
\name{summary_web_objectives}
\alias{summary_web_objectives}
\title{Web Objectives Summary}
\usage{
summary_web_objectives(data, objectives)
}
\arguments{
\item{data;}{dataframe with web data}

\item{objectives;}{labeled vector of URLs}
}
\value{
dataframe
}
\description{
Web Objectives Summary
}

