% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-take_screenshot.R
\name{take_screenshot}
\alias{take_screenshot}
\title{take_screenshot(), Takes screenshots of URLs}
\usage{
take_screenshot(urls = "http://www.google.de", width = 1000)
}
\arguments{
\item{width}{a numeric; width of screenshots in pixel. The width must be the width of the browser
viewport during the experiment!}

\item{URLs}{a character vector; URLs must be full qualified containing schema (http or https)}
}
\value{
vector of screenshot file names that where taken and saved in ./screenshots/
}
\description{
Takes screenshots of all URLs passed to the function arguments and saves them in a new folder
screenshots in the working directory. The filename is the url modified by the function
\code{link{url_to_filename}}.
}
\details{
To get precise heatmaps with the function \code{link{plot_brownie}} it is indispensable to
set the width of the screenshots right. It's recommended to evalualte the max.pos in the data
before taking screenshots. Moreover, when executing the funciton \code{link{plot_brownie}} will
use the files saved in the folder screenshots.
}
\examples{
take_screenshot(c("https://www.visualvest.de/", "https://www.visualvest.de/angebot/", "https://www.visualvest.de/team/"), width = 1000)
}

