% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-transform.R
\name{transform.physio}
\alias{transform.physio}
\title{transform.physio()}
\usage{
\method{transform}{physio}(data)
}
\arguments{
\item{data}{a list of class physio}
}
\value{
list
}
\description{
This function transform a list of class physio, doing the following: remove empty input colums; 
calculate time since tracking has started; normalize the sensor data using z-scores.
Output is again a list, where each subject is a  list item.
}

