% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-plot.R
\name{transform.web}
\alias{transform.web}
\title{transform.web, helper function}
\usage{
\method{transform}{web}(data, url, type = "motion", subject = c(1, 2))
}
\arguments{
\item{data}{a list of class web}

\item{url}{a character}

\item{type}{a character, "motion" or "click"}

\item{subject}{a integer vector including the subjects which should be considered for calculation}
}
\value{
dataframe
}
\description{
Function gets called by \code{link{plot_brownie}} to calculate data points to plot on screenshot.
}
\details{
Extracts data for specified subjects, URL and event. If type euqals "motion" than only data for the
Event MouseMotion is used (for type equals "click" only Event MouseButtonPressed is used).
The movement data is computed by the formulars:
\itemize{
 \item{
 X = H.ScrollPosition + X.Position
 }
 \item{
 X = H.ScrollPosition + X.Position
 }
}
}

