% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-screenshot.R
\name{url_to_filename}
\alias{url_to_filename}
\title{url_to_filename(), Converts a URL to a valid filename}
\usage{
url_to_filename(url)
}
\arguments{
\item{url}{a character; full qualified url to modify}
}
\value{
character
}
\description{
Function removes schema, query of URL and replace all "/" by "-".
}
\details{
Function is necessary to provide a valide filename when saving the screenshots you take with
\code{link{take_screenshot}} of all pages propants visited.
}

