% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-initalize_Unisens.R
\name{initalize_unisens}
\alias{initalize_unisens}
\title{initalize_unisens()}
\usage{
initalize_unisens()
}
\arguments{
\item{no}{input}
}
\value{
void
}
\description{
This function inialize Unisense using rJava.
}

