% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio-marker.R
\name{marker}
\alias{marker}
\title{marker(data, subject)}
\usage{
marker(data, subject, path_marker = F)
}
\arguments{
\item{data}{a list of class exp or web}

\item{subject, }{a integer identifies subject for which marker file shall be created}

\item{path_marker, }{a path; if data shall not be save specify F, if yes, specify a path, where the other physio data is located.
For e.g. "*wd*data/physio/". Saved file name is "marker*subjetct*.csv"}
}
\value{
a dataframe
}
\description{
Creates a "Marker File" for a specified subject. It's needed to calculate for e.g. heart reates for each experimental screen.
For unisens, this file needs to be in *wd*/*data*/physio .
}

